Expression 5: Closures
I have implemented the closures in uni_test.cpp file. func is the copy of the closure created by the lambda expression. I bind the closure to a reference so it is not destroyed at the end of the statement in which it is created. I have reserved the string inside the lambda expression.